requirements.txt
uvicorn
fastapi~=0.91.0
psycopg2-binary
environs~=9.5.0

pip~=21.3.1
wheel~=0.37.1
setuptools~=60.2.0
Django~=4.2

pydantic~=1.10.7
python-dotenv~=1.0.0
guid~=0.2.1
SQLAlchemy~=2.0.9
